requirements.txt
Pillow==9.4.0
einops==0.6.1
kornia==0.6.12
matplotlib==3.6.2
numpy==1.22.2
opencv-python-headless==4.7.0.68
opencv-python==4.5.5.64
pandas==1.5.2
scipy==1.6.3
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
tqdm==4.64.1
xformers==0.0.20

git+https://github.com/Parskatt/RoMa.git@7f8b2e455984c82d957e3318b963cf6673ddaba4
git+https://github.com/Parskatt/DeDoDe.git@d49fe2f8a7a15f8959107ee7a851ced6f54e7621
git+https://github.com/Parskatt/micro-bundle-adjustment.git@934eff87efaf4af3f2f79cfba2de7e60ae4ea5a8
